requirements.txt
pydantic
python-dotenv
ipykernel
langchain
langgraph
langchain-openai
langchain-groq
langchain-core
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
faiss-cpu
langchain_google_genai
langchain-huggingface
hf_xet
langchain-pinecone
youtube_search
chromadb
duckduckgo-search
unstructured
unstructured[pdf]
streamlit